Move into PF subdir
